fix type

+++ intersvc/controlnet_predict_model.go
@@ -3,7 +3,7 @@
  * @Date: 2023-08-22 15:52:43
  * @version:
  * @LastEditors: SpenserCai
- * @LastEditTime: 2023-08-22 16:00:42
+ * @LastEditTime: 2023-08-23 13:00:37
  * @Description: file content
  */
 package intersvc
@@ -14,14 +14,14 @@ type ControlnetPredictArgsItem struct {
 	Model         string  `json:"model"`
 	Weight        float64 `json:"weight"`
 	Image         string  `json:"image"`
-	ResizeMode    int     `json:"resize_mode"`
+	ResizeMode    int64   `json:"resize_mode"`
 	LowVRAM       bool    `json:"low_vram"`
-	ProcessorRes  int     `json:"processor_res"`
-	ThresholdA    int     `json:"threshold_a"`
-	ThresholdB    int     `json:"threshold_b"`
+	ProcessorRes  float64 `json:"processor_res"`
+	ThresholdA    float64 `json:"threshold_a"`
+	ThresholdB    float64 `json:"threshold_b"`
 	GuidanceStart float64 `json:"guidance_start"`
 	GuidanceEnd   float64 `json:"guidance_end"`
-	ControlMode   int     `json:"control_mode"`
+	ControlMode   int64   `json:"control_mode"`
 	PixelPerFect  bool    `json:"pixel_perfect"`
 }
